Fix file names in plugin-html
