+++ packages/plugin-html/src/createHTMLPlugin.js
@@ -1,6 +1,6 @@
 import createToolbar from './toolbar';
 import { createBasePlugin } from 'wix-rich-content-common';
-import { Component } from './html-component';
+import { Component } from './HTMLComponent';
 import { HTML_TYPE } from './types';
 
 const createHTMLPlugin = (config = {}) => {
